+++ Release_Notes.md
@@ -24,6 +24,7 @@
 - Converted the class methods to async/await that were missed in the previous release
     - DataAccess.SQLServer\DDirectories.Save
     - DataAccess.SQLServer\DMessages.Messages
+    - DataAccess.SQLServer\DMessages.Save
 
 ## Net_Core Version 6.0.0.0 (2025-05-16)
 ### Deprecated Classes